intelligence/pure-geo/requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
tensorflow-hub>=0.14.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
imageio>=2.31.0
scikit-image>=0.21.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
h5py>=3.9.0

# Geographic and mapping
geopy>=2.3.0
folium>=0.14.0
geopandas>=0.13.0
pyproj>=3.6.0
rasterio>=1.3.0
shapely>=2.0.0

# Image metadata
exifread>=3.0.0
piexif>=1.1.3

# Web scraping and data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# API and web framework
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0

# Development and testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
pre-commit>=3.3.0

# Optional: GPU acceleration
# nvidia-ml-py3>=7.352.0  # Uncomment if using NVIDIA GPUs